% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmics4ImmunAID-package.R
\docType{package}
\name{MultiOmics4ImmunAID-package}
\alias{MultiOmics4ImmunAID}
\alias{MultiOmics4ImmunAID-package}
\title{MultiOmics4ImmunAID: Multi-omics data analysis for Auto-Inflammatory Diseases}
\description{
Multi-omics data analysis for the ImmunAID project
    (H2020): identify potential biomarkers for Auto-Inflammatory Diseases
    (AID).
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

Authors:
\itemize{
  \item Daniel Herrero-Saboya
}

}
\keyword{internal}
