% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot the connection between blocks}
\usage{
plot_network(
  x,
  C = 1 - diag(length(x)),
  title = paste0("Common rows between blocks : ", length(Reduce(intersect, lapply(x,
    row.names)))),
  cex = 1,
  cex_main = 14 * cex,
  cex_point = 3 * cex,
  cex_nodes = 2 * cex,
  color = c("#eee685", "gray60"),
  shape = "square",
  dashes = TRUE,
  nodes = NULL,
  edges = NULL
)
}
\value{
A dataframe with tuples of connected blocks
}
\description{
Plot the connection between blocks
}
