% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Histogram settings}
\usage{
plotHistogram(
  p = NULL,
  df = NULL,
  hjust = 0,
  vjust = 0.5,
  n = 100,
  title = "",
  color = "black",
  color_gradient = c("#FFF5F0", "#99000D"),
  cex = 1
)
}
\arguments{
\item{p}{A ggplot object.}

\item{df}{A dataframe with a column named "order"}

\item{title}{A character string giving a graphic title}

\item{color}{A vector of character giving the colors for the rows}
}
\description{
Default font for a vertical barplot.
}
\examples{
df <- data.frame(x = runif(30), order = 30:1)
library("ggplot2")
p <- ggplot(df, aes(order, x))
plotHistogram(p, df, "This is my title", "red")
# Add colors per levels of a variable
df$color <- rep(c(1, 2, 3), each = 10)
p <- ggplot(df, aes(order, x, fill = color))
plotHistogram(p, df, "Histogram", as.character(df$color))
}
