% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{omics}
\alias{omics}
\alias{blocks}
\alias{clinic}
\alias{behcet}
\alias{blocks2}
\alias{blocks_w_ms}
\alias{chaq}
\alias{clinic_transf}
\alias{diet}
\alias{elisa}
\alias{haq}
\alias{ms}
\alias{ms_prot_code}
\alias{ms_sample_code}
\alias{pedsql}
\alias{pql}
\alias{rna}
\alias{sf36}
\title{Omics}
\format{
An object of class \code{list} of length 2.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 213 rows and 865 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 33 columns.

An object of class \code{list} of length 3.

An object of class \code{list} of length 2.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 61 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 213 rows and 866 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 183 rows and 118 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 137 rows and 50 columns.

An object of class \code{function} of length 1.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1337 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 77 rows and 49 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 77 rows and 49 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 205 rows and 33426 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 120 rows and 75 columns.
}
\usage{
data(blocks)

data(clinic)

data(behcet)

data(blocks2)

data(blocks_w_ms)

data(chaq)

data(clinic_transf)

data(diet)

data(elisa)

data(haq)

data(ms)

data(ms_prot_code)

data(ms_sample_code)

data(pedsql)

data(pql)

data(rna)

data(sf36)
}
\description{
Tables of omics and clinic from the ImmunAID project (H2020) for
Auto-Inflammatory Diseases (AID).
}
\keyword{datasets}
